FinalPushBeforePresentation
Scene change is implemented. Duplicating gamedevscene but renaming it numerically 1, 2 and so on. Using a combo of strings and integers change the scene upon players reaching door. either someone reaches the door and, a boolean is turned on and waits for others to reach the door. (if they take too long they get deleted) or they load the next area.